requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
WTForms==3.1.2
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.3
openpyxl==3.1.2
Werkzeug==2.3.7
SQLAlchemy==2.0.23
alembic==1.13.1
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
greenlet==3.0.3
typing_extensions==4.9.0
Mako==1.3.2